+++ hapi-fhir-jpaserver-subscription/src/main/java/ca/uhn/fhir/jpa/subscription/model/DebeziumMessage.java
@@ -40,7 +40,7 @@ public class DebeziumMessage extends BaseJsonMessage<PayloadPOJO> {
 		return myPayload.getPayload().getResId();
 	}
 
-	public int getResDeletedAt() {
+	public long getResDeletedAt() {
 		return myPayload.getPayload().getResDeletedAt();
 	}
